Proyecto_Gestoria/requirements.txt
# --- Streamlit ---
streamlit>=1.32
python-dotenv>=1.0
python-dateutil>=2.8

# --- OCR ---
PyMuPDF>=1.24                          # fallback local
azure-ai-documentintelligence>=1.0.0   # Azure OCR (recomendado)
azure-core>=1.30

# --- LLM ---
openai>=1.30
numpy>=1.26

# --- Storage ---
supabase>=2.5

# --- Visualización ---
pandas>=2.2.0
plotly>=5.18
requests>=2.31

# --- Modelos de datos ---
pydantic>=2.7
email-validator>=2.0

# --- Integraciones opcionales (real cuando hay credenciales, simulado si no) ---
# google-api-python-client>=2.0
# google-auth>=2.0
twilio>=8.0

fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
